Add more useful methods
Add `setLEDStatesWithX` and `setLEDStatesWithY` methods, which allow one to set all LEDs in some column (with some x-coordinate) or in some row (with some y-coordinate), respectively, to a given state.
Add `setLEDStatesWithPredicate` method, which allows one to set all LEDs whose x- and/or y-coordinates match a given predicate to a given state, and optionally also set all LEDs not matching the given predicate to the negation of the given state.
As an aside, `setLEDStatesWithX` and `setLEDStatesWithY` do not make use of the `setLEDStatesWithPredicate` method and instead use a single `for` loop in order to run in O(n) rather than O(n^2) time, as they only need to loop through a single column/row.